pyproject.toml
[tool.poetry]
name = "sdss-lvmscraper"
version = "0.0.3"
description = "LVM Instrument data scraper"
authors = ["Florian Briegel <briegel@mpia.de>"]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/sdss/lvmscraper"
repository = "https://github.com/sdss/lvmscraper"
documentation = "https://sdss-lvmscraper.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lvmscraper", from = "python" }
]
include = ["python/lvmscraper/etc/*"]

#[tool.poetry.build]
#script = "build.py"
#generate-setup-file = false

[tool.poetry.scripts]
lvmscraper = "lvmscraper.__main__:scraper"
container_build ="container:build"
container_start ="container:start"
container_stop ="container:stop"
container_isRunning ="container:isRunning"

[tool.poetry.dependencies]
python = ">=3.8,<4"
sdss-cluplus = ">=0.0.18"
CherryPy = ">=18.6.1"
pandas = ">=1.4.0"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = "^3.3.1"
Sphinx = "^3.0.0"
sphinx-autobuild = "^2021.3.14"
sphinx-bootstrap-theme = ">=0.8.0" 
sphinx-markdown-builder = ">=0.5.4"


[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "lvmscraper"
known_sdss = ["sdsstools"]
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
addopts = "--cov lvmscraper --cov-report xml --cov-report html --cov-report term"
testpaths = [
    "tests",
]

[tool.coverage.run]
branch = true
include = ["python/lvmscraper/*"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=49.6.0"]
build-backend = "poetry.core.masonry.api"
